refactoring:  move tests in the separate folder
